Fixed pagination for analog channels and digital channels. Read block length from block header for both analog and digital channels. Added use of WAV:POIN to specify block length as suggested in Siglent code exmaples.

+++ src/hardware/siglent-sds/protocol.h
@@ -126,7 +126,7 @@ struct dev_context {
 	gboolean la_enabled;
 	gboolean channels_switched;
 	double timebase;
-	double max_points;
+	int max_points;
 	float attenuation[MAX_ANALOG_CHANNELS];
 	float vdiv[MAX_ANALOG_CHANNELS];
 	int vert_reference[MAX_ANALOG_CHANNELS];
